refactor: extract bundle stream ownership
